update: Fix scalar test

+++ py-polars/tests/unit/operations/arithmetic/test_list.py
@@ -1043,7 +1043,7 @@ def test_list_boolean_arithmetic_23146() -> None:
     # Boolean list + Scalar
     result = pl.select(pl.lit([True, False]) + 1)
     expected = pl.DataFrame(
-        {"literal": [[2, 1]]}, schema={"literal": pl.List(pl.UInt8)}
+        {"literal": [[2, 1]]}, schema={"literal": pl.List(pl.UInt32)}
     )
     assert_frame_equal(result, expected)
